refactor(core): [WIP] continue preflight refactor
